Move notification emails logic to a separate script
